fix(seo): scope guides index structured data to the index page
The /guides segment layout rendered the all-guides ItemList and a partial
"Home > Guides" BreadcrumbList. Because a layout wraps every child route,
that schema leaked onto every individual guide page, giving each one two
conflicting BreadcrumbList trails plus an irrelevant ItemList — the likely
reason breadcrumb rich results never triggered in Search Console.

Move that index-only schema into the guides index page.tsx, rendered via a
new shared, typed, XSS-safe <JsonLd> component. Each guide detail page now
emits exactly one Home > Guides > [Title] BreadcrumbList alongside its
existing Article schema.
